Update Iris Dataset.py
Introduction:
This is perhaps the best known database to be found in the pattern recognition literature. 
Fisher's paper is a classic in the field and is referenced frequently to this day. (See Duda & Hart, for example.) 
The data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant. 
One class is linearly separable from the other 2; the latter are not linearly separable from each other.

Attribute Information:

1. sepal length in cm
2. sepal width in cm
3. petal length in cm
4. petal width in cm
5. class:
1) Iris Setosa
2) Iris Versicolour
3) Iris Virginica

Objective - Identifying the best model for the dataset

Table of Content
- Importing the Relevant Libraries
- Importing the Dataset
- Data Inspection
- Exploratory Data Analysis
- Making a Prediction
- Comparison of the results of the models